requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.1.0
bleach==1.5.0
certifi==2019.11.28
cleverhans==3.0.1
cloudpickle==1.2.2
colorama==0.4.1
cycler==0.10.0
Cython==0.29.14
cytoolz==0.10.1
dask==2.9.0
decorator==4.4.1
gast==0.3.2
grpcio==1.26.0
h5py==2.8.0
html5lib==0.9999999
imageio==2.6.1
ipython==7.10.1
ipython-genutils==0.2.0
jedi==0.15.1
joblib==0.14.1
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mnist==0.2.2
networkx==2.4
nose==1.3.7
numpy==1.17.4
olefile==0.46
opencv-python==4.1.2.30
pandas==0.25.3
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==3.0.2
protobuf==3.5.2
pycodestyle==2.5.0
pydotplus==2.0.2
Pygments==2.5.2
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
scikit-image==0.15.0
scikit-learn==0.22.1
scipy==1.1.0
six==1.13.0
sklearn==0.0
tensorboard==1.7.0
tensorflow==1.7.0
tensorflow-probability==0.9.0
termcolor==1.1.0
toolz==0.10.0
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
Werkzeug==0.16.0
wincertstore==0.2
